requirements.txt
discord
python-dotenv>=0.15.0
PyYAML>=5.3.1
pymongo
